requirements.txt
# Core
fastmcp>=0.4.0
mcp>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# APIs
amadeus>=8.0.0
google-generativeai>=0.3.0

# Interface
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
